seeder/requirements.txt
brightdata-sdk>=2.2.0
openai>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
